Always use correct list of files.
If you reference a local clone of this gem from another by overriding Bundler's lookups (e.g. adding this to `~/.bundle/config`):

```ruby
BUNDLE_LOCAL__FLUENT___PLUGIN___DOGSTATSD: "/Users/somebody/somewhere/fluent-plugin-dogstatsd"
```

The `git ls-files` in the `.gemspec` file is executed in the working directory of the project referencing this one.

Ensuring we're in the correct directory probably shouldn't be needed, but adding this guard code should eliminate an admittedly small, but still annoying, sharp corner.